requirements.txt
tensorflow>=2.10.0
numpy
matplotlib
flask
streamlit
Pillow
requests
